Remove obsolete comments, tweak others
There are a few comments here that were very, very old and predate
dbfawk.

One is no longer true.

The other refers to how a graphics context used to be used for beam
headings, but it was put in in 2003.

Since we haven't been using that graphics context for that purpose in
22 years, why continue saying we used to do it?

Tidied up some other comments, tweaking language or formatting, and
removing some that were put there during development, apparently
for ease of finding where work was ongoing.

The Xastir code formatting rules also say we're supposed to have four
lines between function definitions.  I had not respected that in all
my hackage on map_shp.c, but I'm fixing that now.