Make tiny cosmetic changes
Change name of "index" variable to "numXPoints" because that is all it
was being used for after recent refactors instead of being a generic
loop variable *and* other things --- it's the number of points that
have been added to the XPoint array.

Change one use of the variable "i" to use of "numXPoints" because that
is what it was being used for in that place, and was no longer a loop
index in that spot.

Change the code that plots shapefile SHPT_POINT type to call the
"symbol" function in one place instead of two.

Reformat comments that had previously been deeply indented and
therefore had short lines.  They're not indented so deeply now, and
that looked funny.

Remove the "high_water_mark_i" variable, which was used for debugging
of shapefiles back in 2002 when the XPoint array was being overrun in
one place.  It is now redundant with the variable
high_water_mark_index, which can still be used to determine when we've
been given a shapefile with a part so huge we can't store it without
overrunning the array, and doing so will still cause an error message
to be printed.  Change that error message to be more generic, it had
previously been cut/pasted from the SHPT_ARC code but is now shared by
SHPT_ARC and SHPT_POLYGON.

Reformat code lines that no longer need to be so short due to
indentation changes.

Add a couple of comments explaining code that I briefly considered
"cleaning up" only to realize that doing so would change
behavior (notably, we need to compute polygon holes whether or not we
actually do filling, because we draw them differently from polygon
boundaries if we're not filling polygons).

Remove use of an unnecessary temporary variable. That usage is a relic
of pre-dbfawk code, and at this time it was just copying a
non-temporary variable and using it.  Just use the non-temp there now,
which slightly simplifies the conditional the code's in as well.