Move initialization of rendering variables to a function
All of the static variables that draw_shapefile_map uses must be
reinitialized every time we're called, otherwise a dbfawk file that
forgets to set some of them winds up getting the values that the
previous map used --- making the rendering of that second map
dependent on which map was displayed before it.

Move that tiny block of variable initialization into a function, for
the sole purpose of having a top level function that is more
easily read by an unfamiliar developer.

This pertains more to #268 than to #274, but at this point I'm having
trouble keeping the lines between those two issues distinct.