Make sure testdbfawk agree with map_shp.c
testdbfawk, a utility for checking dbfawk files against the shapefiles
to which they pertain, only read and printed a subset of the variables
that map_shp actually cares about.

This commit adds those variables to testdbfawk, makes it add them to
the symbol table it uses, and makes it print what it finds for them.

This is not really related to the issue I'm working on now (#274), but the
issues that inspired working on the issue I'm working on now (#152, #157,
and maybe even #32 should that be possible).  But it's so minor I
don't feel bad about including it in this branch.