Add colors, fix bug in multipart polygon stippling
In my investigations of the Natural Earth countries shapefile, I was
puzzled that a large number of countries weren't getting stippled,
even though my dbfawk file said to stipple everything.

The reason was that the draw_shapefile_map code resets the fill style
to solid right before it draws labels, and was only setting the fill
style to stippled before the loop over shape parts.

Thus, only the first part of a multipart shape would be stippled, all
subsequent parts would be solid.

This commit moves the setting of the polygon stipple type to right
before each filled part is drawn.

I have also discovered that the Natural Earth shapefiles have so many
attributes that dbfawk puked on them because it has a hard-coded
length of the "in" array of 1024 bytes, so it reported the line was
too long.   I have increased the size of the "in" array to 2048.

Finally, I added 18 new colors to xastir.rgb, and loaded them into the
colors array in main.c.  This required increasing the MAX_COLORS macro
in colors.h.

The colors I added are now recommended as standard map palattes by
some sources, and I am trying out the "Set1" colors for the Natural
Earth shapefiles.

Related to #152