Add another place for the stipple to be set
In a recent commit I moved the call to set_shpt_polygon_fill_stipple
to right before the filled polygon was drawn, so that it would apply
to every part of that shape.  Prior to that, the Natural Earth countries
rendered really weird.

But I found on zooming in that some of the very old shapefiles I had
stopped rendering correctly.  These all had only ONE part in every
shape, and managed never to be stippled at all.

So I have replaced the call to set_shpt_polygon_fill_stipple that I
had removed at the beginning of the loop, so now it gets called
there *and* every time a new part is drawn.

This is necessary because the stipple needs to be set before we set
up the graphics context for polygon holes, otherwise the fill style of
the clip mask winds up being solid.