Add a new polygon label location selection method
This commit adds three new dbfawk variables:
  - label_method
  - label_lon
  - label_lat

If label_method is set in the dbfawk file to 1 and a shape has set
label_lon and label_lat to something nonzero, Xastir will now use
label_lat and label_lon as the point at which to draw a polygon's
label.

If label_method is anything other than 1, or if either lat/lon is
exactly 0.0, we use the label location selection method we've always
used, the center of the shape's bounding box as stored in the
shapefile object.

Closes #289.