Add a min_display_level variable to dbfawk
Prior to this commit, if you wanted a shapefile layer to be hidden at
low zoom levels, you had to set the Min Zoom value in the map
properties.  That applies to the entire file.

This commit adds a "min_display_level" variable to those recognized by
dbfawk and draw_shapefile_map.  If the zoom level is smaller than
min_display_level, don't show the shape.  This would allow
finer-grained control over map rendering as well as file-wide skipping.

This is especially helpful for large (worldwide) shapefiles with fill,
because when zoomed in tight you don't generally want your map shaded
with the fill of the country you're in.

But it will also enable me to make use of the Natural Earth Vector's
map-rendering-friendly DBF fields, which also include things like
"max_zoom" and "min_zoom" to help map makers decide which shapes of a
shapefile to display at various zooms.

Closes #294