Make Xastir choose an appropriate starting map
If a user is starting Xastir for the very first time, without ever
having had a .xastir directory and configuration, Xastir tries to give
that user a map to look at.  To date, that has been the "worldhi.map".

This commit changes that code so that if the build has shapefile
support compiled in, we select the NaturalEarth directory instead.

If shapefile support is NOT compiled in, we select worldhi.map.

This initial selection is done by writing to the selected_maps.sys
file in the newly-created .xastir directory.

Closes #152