Replace hard-coded screen coordinates limits
draw_shapefile_map has, for a long time, had code that used hard-coded
values of "1700l" to declare points in a shapefile as having "wacko"
screen coordinates --- bigger than 1700l, they got reset to 1700l.

But the thing this code was trying to do was to make sure that
coordinates were not so huge that they were way off screen.

On large screens (such as mine), this clipped things off well inside
the screen boundaries.  Notably, when these limits were applied to the
clipping region used to handle polygon shapefiles with holes and
those polygons were much bigger than the screen, the clipping mask
would clip them off inside the screen.

This commit fixes that.  Instead of using a hard-coded 1700l (which
was probably appropriate back in 2003 when this code was written, and
monitors of width greater than 1700 weren't common), it uses the
global variables screen_width and screen_height.

closes #288