Add dbfawk file for 2025 edition Tiger/Line roads
First of many, this commit adds a rendering hints file for the 2025
edition Tiger/Line shapefiles for roads.  It can be used either with
the county-level "all roads" files or the state-level "prisecroads"
shapefile that has only primary and secondary roads.  Both have the
same signature and therefore share the same rendering rules.