Replace the outdated TIGER/Line address lookup with Nominatim based lookup (#301)
This PR replaces the geocoding lookup using processed TIGER/Line data with lookups from OpenStreetMaps using Nominatim. The TIGER/Line method relies on out of date data.

The Maps->Find Address menu item has been replaced with a Maps->Search Location item and adds a Maps->Configure->Geocoding Settings menu.
The search is free form and supports a filter based on the country.

It should be possible to run Nominatim locally but I have not tried that.

This feature depends on libcurl and the cJSON library. Configure has been updated to account for this.

A future PR is still needed to completely remove the old TIGER/Line geocoding code.