Add colors, refine tl_2025 road rendering
Now that I have the USTigermap.geo working again, I'm trying to make
the road layer from the Tiger shapefiles look as much like the online
one as possible.

This commit changes the colors of primary, secondary, and local roads
to be like those in the WMS Tigermap.  The biggest difference is that
I've made the local roads a slightly darker shade of gray, because it
needs to stand out against the default gray background Xastir uses.