Add area landmarks file, increase max map points
We had a limit to the number of points in a polygon of
MAX_MAP_POINTS (set in maps.h) of 100k.  Turns out that at
least *some* tiger/line area landmarks data has more than that, and I
was getting warnings about exceeding the array size (the code handles
this by discarding points to prevent actually walking past the array
length).

I think we had this set so small to accommodate old machines with
little memory.  I upped it to 500k, which seems to be perfectly
reasonable these days.

Otherwise, the main point of this commit is to add a new dbfawk file
for Tiger/Line data for area landmarks.