Add dbfawk for tl_2025_us_uac20 shapefile
This is the Urban Areas and Cities shapefile, and it's a single file
covering the entire US.  All features in the file have the same
feature classification code, so they're all just rendered in color
115 ("tgr_city_1") with 13% stippling.

Because other Tiger/Line shapefiles place features inside cities, it
is best that users set a fairly low "layer" number in the Map
Properties dialog for this shapefile, so other layers are drawn on top
of them.  Otherwise, the stippling for the entire city covers up
interior features that are probably more interesting than the city
shading.