Make xastir_udp_client work on FreeBSD
The xastir_udp_client program was throwing "Unable to lookup addresses
for host localhost port 2023" when I tested it out on my BSD machine.
I know I have used this program on this machine before, but not in a
long time, so some OS update must have busted it.

I tracked it down to FreeBSD's getaddrinfo rejecting the "AI_DEFAULT"
flag, which is defined differently on BSD than on some linux systems,
apparently.  The code compiled fine, but doesn't work.

Replacing AI_DEFAULT with (AI_V4MAPPED|AI_ADDRCONFIG) fixed it on my
system.  That is how we were defining AI_DEFAULT if the system didn't
define it for us (a hack added in commit b92956 because at the time
Ubuntu 14.04 didn't define AI_DEFAULT, and some other systems didn't
either).

This commit also adds output of the actual error message from
getaddrinfo using gai_strerror, which is a POSIX standard so should be
available on all of our supported systems.

Closes #312