Add minimum automake version to AM_INIT_AUTOMAKE
Using "$(var)" in a "_SOURCES" variable was broken in versions of
Automake prior to 1.16.  We do that in the test suite.

1.16 came out in 2018, and is probably present in all modern OS
distros, but some old, still operational machines are running older
OSen with old automake, and trying to build Xastir on these machines
will have startling failure when the build hits the test suite.

Putting a minimum version into AM_INIT_AUTOMAKE identifies the problem
at the bootstrap/autoreconf stage and lets the user get a newer
version of automake.

Closes #314