Clean out stale information about GNIS files (#317)
GNIS files ceased to be useful map sources in 2009.  A lengthy
discussion on the xastir@xastir.org mailing list (currently
unavailable for either posting or viewing the archives) ensued because
the map_gnis.c code was changed to recognize the new format, but was
not changed in any way to *use* them properly.  Old format files were
supported by creation of a "map_pop.c" renderer, and old files were
made available on a couple of servers that are long gone.

README.MAPS still documented the state of GNIS use prior to 2009.

This commit removes the get-gnis script, the split_gnis.bash script,
and the split_gnis.pl script.

get-gnis tried to download post-2009 GNIS files from a server that
doesn't exist anymore (the base URL redirects elsewhere, where one can
find ONE file that Xastir can use).  The two split_gnis scripts were
designed to break out state-level GNIS gazetteer files into
county-level files to allow them to be usable on 2004 vintage
computers.  They relied on the old, pre-2009 format of GNIS files to
work, and would be worthless on any files from after 2009.  The slow
map rendering that was of great concern in 2004 is less of an issue
now.

Closes #316