Change actions to get Ubuntu dependencies just so
We are for some reason installing "pcre3", which is actually the
legacy PCRE version (go figure).  The version we really want is pcre2.

PCRE2 appears to be installed on the Ubuntu runner by default, and
configure will pick it over the legacy PCRE if it is installed ---
which it is doing.

Anyhow, as part of going over the Ubuntu-specific install directions
on the wiki, I want to make sure that the stuff we are saying there
matches what we're actually doing.