Clean out nws*.dbfawks
Our config directory was full of ancient nws*.dbfawk files that have
been preserved over the years to maintain backward compatibility with
every variant of NWS shapefile that has come over the internet since
about 2006.

There were so many of them it was hard to tell which was which, and
which was even the one being used by current shapefiles.

I have removed all nws*.dbfawk files that aren't used for the
currently available NWS shapefiles, and renamed the ones that remained
to be "nws?_ddmmyy.dbfawk" with no implied year of last update.

We were inconsistent about whether we moved an old one to a new name
and kept the unadorned name as the current, or if we just created a
new one for every change NWS made to shapefile signatures.

We did this because back in the early 2000s it was onerous to download
whole new sets of shapefiles (slow internet) and many Xastir users
didn't want to do it.  So we kept dbfawks around that would allow
these users to keep their old NWS shapefiles.

But having tons of useless old dbfawks in the directory is a headache,
and these days there's no real need for us to preserve backward
compatibility to 20 year old NWS shapefiles (or even 10 year old
shapefiles).

I have tested that every single one of the current crop of NWS
shapefiles has a DBFAWK file associated with it, but that there's only
one for each type (c, w, z, fz, oz, mz), and they all have a generic
name.

Should NWS change shapefile signatures again, we should not get back
in the habit of supporting old versions just so folks don't have to
bother downloading the current ones.

Closes #303