Add "get-NWSdata" to build
One of the things we really ought to make sure of is that get-NWSdata
actually works and that the files it downloads have dbfawk files that
match signatures.

This commit adds one such test to the CI build, only for the full
Ubuntu build.  It is done right after the make install.