Update README.GIT to a markdown file, expand and clean
README.GIT was a straight-text readme.

Furthermore, it was an early attempt to explain git to developers who
were used to CVS.

I've turned it into markdown so it's nicer to look at on github, and
cleaned out some unnecessary commentary.

In a few places, README.GIT told readers to look at README.GIT for
more detail.  Huh?