Update Makefile.am and README
We have to keep a README around for some reason I can't recall (either
because automake complains about it, or github does) even though we
have a README.md.  We've been doing that by having README be an exact
copy of README.md.  But it was out of sync.  Update it.

Also, Makefile.am is set up to install README.GIT to the docs
directory.  We don't have README.GIT anymore, we have README.GIT.md