Expunge LSB-related hacks
Xastir went through a lot of work to make the code compatible with the
Linux Standard Build (LSB), but LSB is a dead project since 2015.

Keeping scripts and ifdefs and configure tests all over the place to
support a way of supporting LSB makes no sense and bloats the code and
repo.

Closes #324