Remove README.qt
There was effort to add a Qt port of Xastir.  It was initially done as
a branch in CVS, then moved to a branch in git, then broken out
entirely into a completely separate project on Github (which has not
been developed recently, but may be someday).

README.qt documented how to build this code, but was rendered obsolete
when we broke that port out into another project.  Further, it tells
how to build that code with Qt4, which is past its EOL and mostly gone
from package management systems.

Let's remove this bit of obsolete documentation
Closes #326