Remove RPM-related "spec" files
The Xastir project used to produce RPM installers for Xastir.  It has
not done so in years.

These spec files were used for that purpose, but have become outdated
because they're not getting updated, because they're not being used.

Package maintainers of Xastir in Linux distros that use RPMs need to
make their own spec files and keep them updated.  These should not be
here.

Closes #325